Refactored RemoteClusterStateService into smaller files
