+++ server/src/main/java/org/opensearch/gateway/remote/RemoteClusterStateService.java
@@ -11,13 +11,10 @@ package org.opensearch.gateway.remote;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.message.ParameterizedMessage;
-import org.opensearch.Version;
 import org.opensearch.action.LatchedActionListener;
 import org.opensearch.cluster.ClusterState;
-import org.opensearch.cluster.coordination.CoordinationMetadata;
 import org.opensearch.cluster.metadata.IndexMetadata;
 import org.opensearch.cluster.metadata.Metadata;
-import org.opensearch.cluster.metadata.TemplatesMetadata;
 import org.opensearch.common.CheckedRunnable;
 import org.opensearch.common.Nullable;
 import org.opensearch.common.blobstore.BlobContainer;
@@ -30,44 +27,55 @@ import org.opensearch.common.settings.Settings;
 import org.opensearch.common.unit.TimeValue;
 import org.opensearch.common.util.io.IOUtils;
 import org.opensearch.core.action.ActionListener;
-import org.opensearch.core.xcontent.ToXContent;
 import org.opensearch.gateway.remote.ClusterMetadataManifest.UploadedIndexMetadata;
 import org.opensearch.gateway.remote.ClusterMetadataManifest.UploadedMetadataAttribute;
-import org.opensearch.index.remote.RemoteStoreUtils;
 import org.opensearch.index.translog.transfer.BlobStoreTransferService;
 import org.opensearch.node.Node;
 import org.opensearch.node.remotestore.RemoteStoreNodeAttribute;
 import org.opensearch.repositories.RepositoriesService;
 import org.opensearch.repositories.Repository;
 import org.opensearch.repositories.blobstore.BlobStoreRepository;
-import org.opensearch.repositories.blobstore.ChecksumBlobStoreFormat;
 import org.opensearch.threadpool.ThreadPool;
 
 import java.io.Closeable;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
-import java.util.Base64;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Function;
 import java.util.function.LongSupplier;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
-import static java.util.Objects.requireNonNull;
 import static org.opensearch.gateway.PersistedClusterStateService.SLOW_WRITE_LOGGING_THRESHOLD;
+import static org.opensearch.gateway.remote.RemoteClusterStateUtils.DELIMITER;
+import static org.opensearch.gateway.remote.RemoteClusterStateUtils.RemoteStateTransferException;
+import static org.opensearch.gateway.remote.RemoteClusterStateUtils.UploadedMetadataResults;
+import static org.opensearch.gateway.remote.RemoteClusterStateUtils.clusterUUIDContainer;
+import static org.opensearch.gateway.remote.RemoteClusterStateUtils.getCusterMetadataBasePath;
+import static org.opensearch.gateway.remote.RemoteGlobalMetadataManager.COORDINATION_METADATA;
+import static org.opensearch.gateway.remote.RemoteGlobalMetadataManager.COORDINATION_METADATA_FORMAT;
+import static org.opensearch.gateway.remote.RemoteGlobalMetadataManager.CUSTOM_DELIMITER;
+import static org.opensearch.gateway.remote.RemoteGlobalMetadataManager.CUSTOM_METADATA;
+import static org.opensearch.gateway.remote.RemoteGlobalMetadataManager.CUSTOM_METADATA_FORMAT;
+import static org.opensearch.gateway.remote.RemoteGlobalMetadataManager.GLOBAL_METADATA_FORMAT;
+import static org.opensearch.gateway.remote.RemoteGlobalMetadataManager.GLOBAL_METADATA_PATH_TOKEN;
+import static org.opensearch.gateway.remote.RemoteGlobalMetadataManager.SETTINGS_METADATA_FORMAT;
+import static org.opensearch.gateway.remote.RemoteGlobalMetadataManager.SETTING_METADATA;
+import static org.opensearch.gateway.remote.RemoteGlobalMetadataManager.TEMPLATES_METADATA;
+import static org.opensearch.gateway.remote.RemoteGlobalMetadataManager.TEMPLATES_METADATA_FORMAT;
+import static org.opensearch.gateway.remote.RemoteIndexMetadataManager.INDEX_METADATA_FORMAT;
+import static org.opensearch.gateway.remote.RemoteIndexMetadataManager.INDEX_PATH_TOKEN;
+import static org.opensearch.gateway.remote.RemoteManifestManager.MANIFEST_PATH_TOKEN;
 import static org.opensearch.node.remotestore.RemoteStoreNodeAttribute.isRemoteStoreClusterStateEnabled;
 
 /**
@@ -76,101 +84,9 @@ import static org.opensearch.node.remotestore.RemoteStoreNodeAttribute.isRemoteS
  * @opensearch.internal
  */
 public class RemoteClusterStateService implements Closeable {
-
-    public static final String METADATA_NAME_FORMAT = "%s.dat";
-
-    public static final String METADATA_MANIFEST_NAME_FORMAT = "%s";
-
     public static final int RETAINED_MANIFESTS = 10;
 
-    public static final String DELIMITER = "__";
-    public static final String CUSTOM_DELIMITER = "--";
-
     private static final Logger logger = LogManager.getLogger(RemoteClusterStateService.class);
-
-    public static final TimeValue INDEX_METADATA_UPLOAD_TIMEOUT_DEFAULT = TimeValue.timeValueMillis(20000);
-
-    public static final TimeValue GLOBAL_METADATA_UPLOAD_TIMEOUT_DEFAULT = TimeValue.timeValueMillis(20000);
-
-    public static final TimeValue METADATA_MANIFEST_UPLOAD_TIMEOUT_DEFAULT = TimeValue.timeValueMillis(20000);
-
-    public static final Setting<TimeValue> INDEX_METADATA_UPLOAD_TIMEOUT_SETTING = Setting.timeSetting(
-        "cluster.remote_store.state.index_metadata.upload_timeout",
-        INDEX_METADATA_UPLOAD_TIMEOUT_DEFAULT,
-        Setting.Property.Dynamic,
-        Setting.Property.NodeScope
-    );
-
-    public static final Setting<TimeValue> GLOBAL_METADATA_UPLOAD_TIMEOUT_SETTING = Setting.timeSetting(
-        "cluster.remote_store.state.global_metadata.upload_timeout",
-        GLOBAL_METADATA_UPLOAD_TIMEOUT_DEFAULT,
-        Setting.Property.Dynamic,
-        Setting.Property.NodeScope
-    );
-
-    public static final Setting<TimeValue> METADATA_MANIFEST_UPLOAD_TIMEOUT_SETTING = Setting.timeSetting(
-        "cluster.remote_store.state.metadata_manifest.upload_timeout",
-        METADATA_MANIFEST_UPLOAD_TIMEOUT_DEFAULT,
-        Setting.Property.Dynamic,
-        Setting.Property.NodeScope
-    );
-
-    public static final ChecksumBlobStoreFormat<IndexMetadata> INDEX_METADATA_FORMAT = new ChecksumBlobStoreFormat<>(
-        "index-metadata",
-        METADATA_NAME_FORMAT,
-        IndexMetadata::fromXContent
-    );
-
-    public static final ChecksumBlobStoreFormat<Metadata> GLOBAL_METADATA_FORMAT = new ChecksumBlobStoreFormat<>(
-        "metadata",
-        METADATA_NAME_FORMAT,
-        Metadata::fromXContent
-    );
-
-    public static final ChecksumBlobStoreFormat<CoordinationMetadata> COORDINATION_METADATA_FORMAT = new ChecksumBlobStoreFormat<>(
-        "coordination",
-        METADATA_NAME_FORMAT,
-        CoordinationMetadata::fromXContent
-    );
-
-    public static final ChecksumBlobStoreFormat<Settings> SETTINGS_METADATA_FORMAT = new ChecksumBlobStoreFormat<>(
-        "settings",
-        METADATA_NAME_FORMAT,
-        Settings::fromXContent
-    );
-
-    public static final ChecksumBlobStoreFormat<TemplatesMetadata> TEMPLATES_METADATA_FORMAT = new ChecksumBlobStoreFormat<>(
-        "templates",
-        METADATA_NAME_FORMAT,
-        TemplatesMetadata::fromXContent
-    );
-
-    public static final ChecksumBlobStoreFormat<Metadata.Custom> CUSTOM_METADATA_FORMAT = new ChecksumBlobStoreFormat<>(
-        "custom",
-        METADATA_NAME_FORMAT,
-        Metadata.Custom::fromXContent
-    );
-
-    /**
-     * Manifest format compatible with older codec v0, where codec version was missing.
-     */
-    public static final ChecksumBlobStoreFormat<ClusterMetadataManifest> CLUSTER_METADATA_MANIFEST_FORMAT_V0 =
-        new ChecksumBlobStoreFormat<>("cluster-metadata-manifest", METADATA_MANIFEST_NAME_FORMAT, ClusterMetadataManifest::fromXContentV0);
-
-    /**
-     * Manifest format compatible with older codec v1, where global metadata was missing.
-     */
-    public static final ChecksumBlobStoreFormat<ClusterMetadataManifest> CLUSTER_METADATA_MANIFEST_FORMAT_V1 =
-        new ChecksumBlobStoreFormat<>("cluster-metadata-manifest", METADATA_MANIFEST_NAME_FORMAT, ClusterMetadataManifest::fromXContentV1);
-
-    /**
-     * Manifest format compatible with codec v2, where we introduced codec versions/global metadata.
-     */
-    public static final ChecksumBlobStoreFormat<ClusterMetadataManifest> CLUSTER_METADATA_MANIFEST_FORMAT = new ChecksumBlobStoreFormat<>(
-        "cluster-metadata-manifest",
-        METADATA_MANIFEST_NAME_FORMAT,
-        ClusterMetadataManifest::fromXContent
-    );
 
     /**
      * Used to specify if cluster state metadata should be published to remote store
@@ -182,18 +98,6 @@ public class RemoteClusterStateService implements Closeable {
         Property.Final
     );
 
-    public static final String CLUSTER_STATE_PATH_TOKEN = "cluster-state";
-    public static final String INDEX_PATH_TOKEN = "index";
-    public static final String GLOBAL_METADATA_PATH_TOKEN = "global-metadata";
-    public static final String MANIFEST_PATH_TOKEN = "manifest";
-    public static final String MANIFEST_FILE_PREFIX = "manifest";
-    public static final String METADATA_FILE_PREFIX = "metadata";
-    public static final String COORDINATION_METADATA = "coordination";
-    public static final String SETTING_METADATA = "settings";
-    public static final String TEMPLATES_METADATA = "templates";
-    public static final String CUSTOM_METADATA = "custom";
-    public static final int SPLITED_MANIFEST_FILE_LENGTH = 6; // file name manifest__term__version__C/P__timestamp__codecversion
-
     private final String nodeId;
     private final Supplier<RepositoriesService> repositoriesService;
     private final Settings settings;
@@ -203,24 +107,13 @@ public class RemoteClusterStateService implements Closeable {
     private BlobStoreTransferService blobStoreTransferService;
     private volatile TimeValue slowWriteLoggingThreshold;
 
-    private volatile TimeValue indexMetadataUploadTimeout;
-    private volatile TimeValue globalMetadataUploadTimeout;
-    private volatile TimeValue metadataManifestUploadTimeout;
-
     private final AtomicBoolean deleteStaleMetadataRunning = new AtomicBoolean(false);
     private final RemotePersistenceStats remoteStateStats;
+    private RemoteIndexMetadataManager remoteIndexMetadataManager;
+    private RemoteGlobalMetadataManager remoteGlobalMetadataManager;
+    private RemoteManifestManager remoteManifestManager;
+    private ClusterSettings clusterSettings;
     public static final int INDEX_METADATA_CURRENT_CODEC_VERSION = 1;
-    public static final int MANIFEST_CURRENT_CODEC_VERSION = ClusterMetadataManifest.CODEC_V2;
-    public static final int GLOBAL_METADATA_CURRENT_CODEC_VERSION = 1;
-
-    // ToXContent Params with gateway mode.
-    // We are using gateway context mode to persist all custom metadata.
-    public static final ToXContent.Params FORMAT_PARAMS;
-    static {
-        Map<String, String> params = new HashMap<>(1);
-        params.put(Metadata.CONTEXT_MODE_PARAM, Metadata.CONTEXT_MODE_GATEWAY);
-        FORMAT_PARAMS = new ToXContent.MapParams(params);
-    }
 
     public RemoteClusterStateService(
         String nodeId,
@@ -236,14 +129,9 @@ public class RemoteClusterStateService implements Closeable {
         this.settings = settings;
         this.relativeTimeNanosSupplier = relativeTimeNanosSupplier;
         this.threadpool = threadPool;
+        this.clusterSettings = clusterSettings;
         this.slowWriteLoggingThreshold = clusterSettings.get(SLOW_WRITE_LOGGING_THRESHOLD);
-        this.indexMetadataUploadTimeout = clusterSettings.get(INDEX_METADATA_UPLOAD_TIMEOUT_SETTING);
-        this.globalMetadataUploadTimeout = clusterSettings.get(GLOBAL_METADATA_UPLOAD_TIMEOUT_SETTING);
-        this.metadataManifestUploadTimeout = clusterSettings.get(METADATA_MANIFEST_UPLOAD_TIMEOUT_SETTING);
         clusterSettings.addSettingsUpdateConsumer(SLOW_WRITE_LOGGING_THRESHOLD, this::setSlowWriteLoggingThreshold);
-        clusterSettings.addSettingsUpdateConsumer(INDEX_METADATA_UPLOAD_TIMEOUT_SETTING, this::setIndexMetadataUploadTimeout);
-        clusterSettings.addSettingsUpdateConsumer(GLOBAL_METADATA_UPLOAD_TIMEOUT_SETTING, this::setGlobalMetadataUploadTimeout);
-        clusterSettings.addSettingsUpdateConsumer(METADATA_MANIFEST_UPLOAD_TIMEOUT_SETTING, this::setMetadataManifestUploadTimeout);
         this.remoteStateStats = new RemotePersistenceStats();
     }
 
@@ -276,7 +164,7 @@ public class RemoteClusterStateService implements Closeable {
             true,
             true
         );
-        final ClusterMetadataManifest manifest = uploadManifest(
+        final ClusterMetadataManifest manifest = remoteManifestManager.uploadManifest(
             clusterState,
             uploadedMetadataResults.uploadedIndexMetadata,
             previousClusterUUID,
@@ -345,7 +233,10 @@ public class RemoteClusterStateService implements Closeable {
             clusterState.metadata()
         ) == false;
         final Map<String, UploadedMetadataAttribute> previousStateCustomMap = new HashMap<>(previousManifest.getCustomMetadataMap());
-        final Map<String, Metadata.Custom> customsToUpload = getUpdatedCustoms(clusterState, previousClusterState);
+        final Map<String, Metadata.Custom> customsToUpload = remoteGlobalMetadataManager.getUpdatedCustoms(
+            clusterState,
+            previousClusterState
+        );
         final Map<String, UploadedMetadataAttribute> allUploadedCustomMap = new HashMap<>(previousManifest.getCustomMetadataMap());
         for (final String custom : clusterState.metadata().customs().keySet()) {
             // remove all the customs which are present currently
@@ -407,7 +298,7 @@ public class RemoteClusterStateService implements Closeable {
         // remove the data for removed custom/indices
         previousStateCustomMap.keySet().forEach(allUploadedCustomMap::remove);
         previousStateIndexMetadataVersionByName.keySet().forEach(allUploadedIndexMetadata::remove);
-        final ClusterMetadataManifest manifest = uploadManifest(
+        final ClusterMetadataManifest manifest = remoteManifestManager.uploadManifest(
             clusterState,
             new ArrayList<>(allUploadedIndexMetadata.values()),
             previousManifest.getPreviousClusterUUID(),
@@ -493,7 +384,7 @@ public class RemoteClusterStateService implements Closeable {
         if (uploadSettingsMetadata) {
             uploadTasks.put(
                 SETTING_METADATA,
-                getAsyncMetadataWriteAction(
+                remoteGlobalMetadataManager.getAsyncMetadataWriteAction(
                     clusterState,
                     SETTING_METADATA,
                     SETTINGS_METADATA_FORMAT,
@@ -505,7 +396,7 @@ public class RemoteClusterStateService implements Closeable {
         if (uploadCoordinationMetadata) {
             uploadTasks.put(
                 COORDINATION_METADATA,
-                getAsyncMetadataWriteAction(
+                remoteGlobalMetadataManager.getAsyncMetadataWriteAction(
                     clusterState,
                     COORDINATION_METADATA,
                     COORDINATION_METADATA_FORMAT,
@@ -517,7 +408,7 @@ public class RemoteClusterStateService implements Closeable {
         if (uploadTemplateMetadata) {
             uploadTasks.put(
                 TEMPLATES_METADATA,
-                getAsyncMetadataWriteAction(
+                remoteGlobalMetadataManager.getAsyncMetadataWriteAction(
                     clusterState,
                     TEMPLATES_METADATA,
                     TEMPLATES_METADATA_FORMAT,
@@ -530,11 +421,20 @@ public class RemoteClusterStateService implements Closeable {
             String customComponent = String.join(CUSTOM_DELIMITER, CUSTOM_METADATA, key);
             uploadTasks.put(
                 customComponent,
-                getAsyncMetadataWriteAction(clusterState, customComponent, CUSTOM_METADATA_FORMAT, value, listener)
+                remoteGlobalMetadataManager.getAsyncMetadataWriteAction(
+                    clusterState,
+                    customComponent,
+                    CUSTOM_METADATA_FORMAT,
+                    value,
+                    listener
+                )
             );
         });
         indexToUpload.forEach(indexMetadata -> {
-            uploadTasks.put(indexMetadata.getIndexName(), getIndexMetadataAsyncAction(clusterState, indexMetadata, listener));
+            uploadTasks.put(
+                indexMetadata.getIndexName(),
+                remoteIndexMetadataManager.getIndexMetadataAsyncAction(clusterState, indexMetadata, listener)
+            );
         });
 
         // start async upload of all required metadata files
@@ -543,7 +443,7 @@ public class RemoteClusterStateService implements Closeable {
         }
 
         try {
-            if (latch.await(getGlobalMetadataUploadTimeout().millis(), TimeUnit.MILLISECONDS) == false) {
+            if (latch.await(remoteGlobalMetadataManager.getGlobalMetadataUploadTimeout().millis(), TimeUnit.MILLISECONDS) == false) {
                 // TODO: We should add metrics where transfer is timing out. [Issue: #10687]
                 RemoteStateTransferException ex = new RemoteStateTransferException(
                     String.format(
@@ -603,75 +503,6 @@ public class RemoteClusterStateService implements Closeable {
         return response;
     }
 
-    /**
-     * Allows async Upload of IndexMetadata to remote
-     *
-     * @param clusterState current ClusterState
-     * @param indexMetadata {@link IndexMetadata} to upload
-     * @param latchedActionListener listener to respond back on after upload finishes
-     */
-    private CheckedRunnable<IOException> getIndexMetadataAsyncAction(
-        ClusterState clusterState,
-        IndexMetadata indexMetadata,
-        LatchedActionListener<ClusterMetadataManifest.UploadedMetadata> latchedActionListener
-    ) {
-        final BlobContainer indexMetadataContainer = indexMetadataContainer(
-            clusterState.getClusterName().value(),
-            clusterState.metadata().clusterUUID(),
-            indexMetadata.getIndexUUID()
-        );
-        final String indexMetadataFilename = indexMetadataFileName(indexMetadata);
-        ActionListener<Void> completionListener = ActionListener.wrap(
-            resp -> latchedActionListener.onResponse(
-                new UploadedIndexMetadata(
-                    indexMetadata.getIndex().getName(),
-                    indexMetadata.getIndexUUID(),
-                    indexMetadataContainer.path().buildAsString() + indexMetadataFilename
-                )
-            ),
-            ex -> latchedActionListener.onFailure(new RemoteStateTransferException(indexMetadata.getIndex().toString(), ex))
-        );
-
-        return () -> INDEX_METADATA_FORMAT.writeAsyncWithUrgentPriority(
-            indexMetadata,
-            indexMetadataContainer,
-            indexMetadataFilename,
-            blobStoreRepository.getCompressor(),
-            completionListener,
-            FORMAT_PARAMS
-        );
-    }
-
-    /**
-     * Allows async upload of Metadata components to remote
-     */
-
-    private CheckedRunnable<IOException> getAsyncMetadataWriteAction(
-        ClusterState clusterState,
-        String component,
-        ChecksumBlobStoreFormat componentMetadataBlobStore,
-        ToXContent componentMetadata,
-        LatchedActionListener<ClusterMetadataManifest.UploadedMetadata> latchedActionListener
-    ) {
-        final BlobContainer globalMetadataContainer = globalMetadataContainer(
-            clusterState.getClusterName().value(),
-            clusterState.metadata().clusterUUID()
-        );
-        final String componentMetadataFilename = metadataAttributeFileName(component, clusterState.metadata().version());
-        ActionListener<Void> completionListener = ActionListener.wrap(
-            resp -> latchedActionListener.onResponse(new UploadedMetadataAttribute(component, componentMetadataFilename)),
-            ex -> latchedActionListener.onFailure(new RemoteStateTransferException(component, ex))
-        );
-        return () -> componentMetadataBlobStore.writeAsyncWithUrgentPriority(
-            componentMetadata,
-            globalMetadataContainer,
-            componentMetadataFilename,
-            blobStoreRepository.getCompressor(),
-            completionListener,
-            FORMAT_PARAMS
-        );
-    }
-
     @Nullable
     public ClusterMetadataManifest markLastStateAsCommitted(ClusterState clusterState, ClusterMetadataManifest previousManifest)
         throws IOException {
@@ -681,7 +512,7 @@ public class RemoteClusterStateService implements Closeable {
             return null;
         }
         assert previousManifest != null : "Last cluster metadata manifest is not set";
-        ClusterMetadataManifest committedManifest = uploadManifest(
+        ClusterMetadataManifest committedManifest = remoteManifestManager.uploadManifest(
             clusterState,
             previousManifest.getIndices(),
             previousManifest.getPreviousClusterUUID(),
@@ -693,6 +524,17 @@ public class RemoteClusterStateService implements Closeable {
         );
         deleteStaleClusterUUIDs(clusterState, committedManifest);
         return committedManifest;
+    }
+
+    /**
+     * Fetch latest ClusterMetadataManifest from remote state store
+     *
+     * @param clusterUUID uuid of cluster state to refer to in remote
+     * @param clusterName name of the cluster
+     * @return ClusterMetadataManifest
+     */
+    public Optional<ClusterMetadataManifest> getLatestClusterMetadataManifest(String clusterName, String clusterUUID) {
+        return remoteManifestManager.getLatestClusterMetadataManifest(clusterName, clusterUUID);
     }
 
     @Override
@@ -711,131 +553,16 @@ public class RemoteClusterStateService implements Closeable {
         final Repository repository = repositoriesService.get().repository(remoteStoreRepo);
         assert repository instanceof BlobStoreRepository : "Repository should be instance of BlobStoreRepository";
         blobStoreRepository = (BlobStoreRepository) repository;
-    }
-
-    private ClusterMetadataManifest uploadV1Manifest(
-        ClusterState clusterState,
-        List<UploadedIndexMetadata> uploadedIndexMetadata,
-        String previousClusterUUID,
-        String globalMetadataFileName,
-        boolean committed
-    ) throws IOException {
-        synchronized (this) {
-            final String manifestFileName = getManifestFileName(
-                clusterState.term(),
-                clusterState.version(),
-                committed,
-                ClusterMetadataManifest.CODEC_V1
-            );
-            ClusterMetadataManifest manifest = ClusterMetadataManifest.builder()
-                .clusterTerm(clusterState.term())
-                .stateVersion(clusterState.getVersion())
-                .clusterUUID(clusterState.metadata().clusterUUID())
-                .stateUUID(clusterState.stateUUID())
-                .opensearchVersion(Version.CURRENT)
-                .nodeId(nodeId)
-                .committed(committed)
-                .codecVersion(ClusterMetadataManifest.CODEC_V1)
-                .globalMetadataFileName(globalMetadataFileName)
-                .indices(uploadedIndexMetadata)
-                .previousClusterUUID(previousClusterUUID)
-                .clusterUUIDCommitted(clusterState.metadata().clusterUUIDCommitted())
-                .build();
-            writeMetadataManifest(clusterState.getClusterName().value(), clusterState.metadata().clusterUUID(), manifest, manifestFileName);
-            return manifest;
-        }
-    }
-
-    private ClusterMetadataManifest uploadManifest(
-        ClusterState clusterState,
-        List<UploadedIndexMetadata> uploadedIndexMetadata,
-        String previousClusterUUID,
-        UploadedMetadataAttribute uploadedCoordinationMetadata,
-        UploadedMetadataAttribute uploadedSettingsMetadata,
-        UploadedMetadataAttribute uploadedTemplatesMetadata,
-        Map<String, UploadedMetadataAttribute> uploadedCustomMetadataMap,
-        boolean committed
-    ) throws IOException {
-        synchronized (this) {
-            final String manifestFileName = getManifestFileName(
-                clusterState.term(),
-                clusterState.version(),
-                committed,
-                MANIFEST_CURRENT_CODEC_VERSION
-            );
-            final ClusterMetadataManifest manifest = new ClusterMetadataManifest(
-                clusterState.term(),
-                clusterState.getVersion(),
-                clusterState.metadata().clusterUUID(),
-                clusterState.stateUUID(),
-                Version.CURRENT,
-                nodeId,
-                committed,
-                MANIFEST_CURRENT_CODEC_VERSION,
-                null,
-                uploadedIndexMetadata,
-                previousClusterUUID,
-                clusterState.metadata().clusterUUIDCommitted(),
-                uploadedCoordinationMetadata,
-                uploadedSettingsMetadata,
-                uploadedTemplatesMetadata,
-                uploadedCustomMetadataMap
-            );
-            writeMetadataManifest(clusterState.getClusterName().value(), clusterState.metadata().clusterUUID(), manifest, manifestFileName);
-            return manifest;
-        }
-    }
-
-    private void writeMetadataManifest(String clusterName, String clusterUUID, ClusterMetadataManifest uploadManifest, String fileName)
-        throws IOException {
-        AtomicReference<String> result = new AtomicReference<String>();
-        AtomicReference<Exception> exceptionReference = new AtomicReference<Exception>();
-
-        final BlobContainer metadataManifestContainer = manifestContainer(clusterName, clusterUUID);
-
-        // latch to wait until upload is not finished
-        CountDownLatch latch = new CountDownLatch(1);
-
-        LatchedActionListener completionListener = new LatchedActionListener<>(ActionListener.wrap(resp -> {
-            logger.trace(String.format(Locale.ROOT, "Manifest file uploaded successfully."));
-        }, ex -> { exceptionReference.set(ex); }), latch);
-
-        getClusterMetadataManifestBlobStoreFormat(fileName).writeAsyncWithUrgentPriority(
-            uploadManifest,
-            metadataManifestContainer,
-            fileName,
-            blobStoreRepository.getCompressor(),
-            completionListener,
-            FORMAT_PARAMS
-        );
-
-        try {
-            if (latch.await(getMetadataManifestUploadTimeout().millis(), TimeUnit.MILLISECONDS) == false) {
-                RemoteStateTransferException ex = new RemoteStateTransferException(
-                    String.format(Locale.ROOT, "Timed out waiting for transfer of manifest file to complete")
-                );
-                throw ex;
-            }
-        } catch (InterruptedException ex) {
-            RemoteStateTransferException exception = new RemoteStateTransferException(
-                String.format(Locale.ROOT, "Timed out waiting for transfer of manifest file to complete - %s"),
-                ex
-            );
-            Thread.currentThread().interrupt();
-            throw exception;
-        }
-        if (exceptionReference.get() != null) {
-            throw new RemoteStateTransferException(exceptionReference.get().getMessage(), exceptionReference.get());
-        }
-        logger.debug(
-            "Metadata manifest file [{}] written during [{}] phase. ",
-            fileName,
-            uploadManifest.isCommitted() ? "commit" : "publish"
-        );
+        remoteGlobalMetadataManager = new RemoteGlobalMetadataManager(blobStoreRepository, clusterSettings);
+        remoteIndexMetadataManager = new RemoteIndexMetadataManager(blobStoreRepository, clusterSettings);
+        remoteManifestManager = new RemoteManifestManager(blobStoreRepository, clusterSettings, nodeId);
     }
 
     private String fetchPreviousClusterUUID(String clusterName, String clusterUUID) {
-        final Optional<ClusterMetadataManifest> latestManifest = getLatestClusterMetadataManifest(clusterName, clusterUUID);
+        final Optional<ClusterMetadataManifest> latestManifest = remoteManifestManager.getLatestClusterMetadataManifest(
+            clusterName,
+            clusterUUID
+        );
         if (!latestManifest.isPresent()) {
             final String previousClusterUUID = getLastKnownUUIDFromRemote(clusterName);
             assert !clusterUUID.equals(previousClusterUUID) : "Last cluster UUID is same current cluster UUID";
@@ -844,188 +571,8 @@ public class RemoteClusterStateService implements Closeable {
         return latestManifest.get().getPreviousClusterUUID();
     }
 
-    private BlobContainer indexMetadataContainer(String clusterName, String clusterUUID, String indexUUID) {
-        // 123456789012_test-cluster/cluster-state/dsgYj10Nkso7/index/ftqsCnn9TgOX
-        return blobStoreRepository.blobStore()
-            .blobContainer(getCusterMetadataBasePath(clusterName, clusterUUID).add(INDEX_PATH_TOKEN).add(indexUUID));
-    }
-
-    private BlobContainer globalMetadataContainer(String clusterName, String clusterUUID) {
-        // 123456789012_test-cluster/cluster-state/dsgYj10Nkso7/global-metadata/
-        return blobStoreRepository.blobStore()
-            .blobContainer(getCusterMetadataBasePath(clusterName, clusterUUID).add(GLOBAL_METADATA_PATH_TOKEN));
-    }
-
-    private BlobContainer manifestContainer(String clusterName, String clusterUUID) {
-        // 123456789012_test-cluster/cluster-state/dsgYj10Nkso7/manifest
-        return blobStoreRepository.blobStore().blobContainer(getManifestFolderPath(clusterName, clusterUUID));
-    }
-
-    private BlobPath getCusterMetadataBasePath(String clusterName, String clusterUUID) {
-        return blobStoreRepository.basePath().add(encodeString(clusterName)).add(CLUSTER_STATE_PATH_TOKEN).add(clusterUUID);
-    }
-
-    private BlobContainer clusterUUIDContainer(String clusterName) {
-        return blobStoreRepository.blobStore()
-            .blobContainer(
-                blobStoreRepository.basePath()
-                    .add(Base64.getUrlEncoder().withoutPadding().encodeToString(clusterName.getBytes(StandardCharsets.UTF_8)))
-                    .add(CLUSTER_STATE_PATH_TOKEN)
-            );
-    }
-
     private void setSlowWriteLoggingThreshold(TimeValue slowWriteLoggingThreshold) {
         this.slowWriteLoggingThreshold = slowWriteLoggingThreshold;
-    }
-
-    private void setIndexMetadataUploadTimeout(TimeValue newIndexMetadataUploadTimeout) {
-        this.indexMetadataUploadTimeout = newIndexMetadataUploadTimeout;
-    }
-
-    private void setGlobalMetadataUploadTimeout(TimeValue newGlobalMetadataUploadTimeout) {
-        this.globalMetadataUploadTimeout = newGlobalMetadataUploadTimeout;
-    }
-
-    private void setMetadataManifestUploadTimeout(TimeValue newMetadataManifestUploadTimeout) {
-        this.metadataManifestUploadTimeout = newMetadataManifestUploadTimeout;
-    }
-
-    private Map<String, Metadata.Custom> getUpdatedCustoms(ClusterState currentState, ClusterState previousState) {
-        if (Metadata.isCustomMetadataEqual(previousState.metadata(), currentState.metadata())) {
-            return new HashMap<>();
-        }
-        Map<String, Metadata.Custom> updatedCustom = new HashMap<>();
-        Set<String> currentCustoms = new HashSet<>(currentState.metadata().customs().keySet());
-        for (Map.Entry<String, Metadata.Custom> cursor : previousState.metadata().customs().entrySet()) {
-            if (cursor.getValue().context().contains(Metadata.XContentContext.GATEWAY)) {
-                if (currentCustoms.contains(cursor.getKey())
-                    && !cursor.getValue().equals(currentState.metadata().custom(cursor.getKey()))) {
-                    // If the custom metadata is updated, we need to upload the new version.
-                    updatedCustom.put(cursor.getKey(), currentState.metadata().custom(cursor.getKey()));
-                }
-                currentCustoms.remove(cursor.getKey());
-            }
-        }
-        for (String custom : currentCustoms) {
-            Metadata.Custom cursor = currentState.metadata().custom(custom);
-            if (cursor.context().contains(Metadata.XContentContext.GATEWAY)) {
-                updatedCustom.put(custom, cursor);
-            }
-        }
-        return updatedCustom;
-    }
-
-    public TimeValue getIndexMetadataUploadTimeout() {
-        return this.indexMetadataUploadTimeout;
-    }
-
-    public TimeValue getGlobalMetadataUploadTimeout() {
-        return this.globalMetadataUploadTimeout;
-    }
-
-    public TimeValue getMetadataManifestUploadTimeout() {
-        return this.metadataManifestUploadTimeout;
-    }
-
-    static String getManifestFileName(long term, long version, boolean committed, int codecVersion) {
-        // 123456789012_test-cluster/cluster-state/dsgYj10Nkso7/manifest/manifest__<inverted_term>__<inverted_version>__C/P__<inverted__timestamp>__<codec_version>
-        return String.join(
-            DELIMITER,
-            MANIFEST_PATH_TOKEN,
-            RemoteStoreUtils.invertLong(term),
-            RemoteStoreUtils.invertLong(version),
-            (committed ? "C" : "P"), // C for committed and P for published
-            RemoteStoreUtils.invertLong(System.currentTimeMillis()),
-            String.valueOf(codecVersion) // Keep the codec version at last place only, during read we reads last place to
-                                         // determine codec version.
-        );
-    }
-
-    static String indexMetadataFileName(IndexMetadata indexMetadata) {
-        // 123456789012_test-cluster/cluster-state/dsgYj10Nkso7/index/<index_UUID>/metadata__<inverted_index_metadata_version>__<inverted__timestamp>__<codec
-        // version>
-        return String.join(
-            DELIMITER,
-            METADATA_FILE_PREFIX,
-            RemoteStoreUtils.invertLong(indexMetadata.getVersion()),
-            RemoteStoreUtils.invertLong(System.currentTimeMillis()),
-            String.valueOf(INDEX_METADATA_CURRENT_CODEC_VERSION) // Keep the codec version at last place only, during read we reads last
-                                                                 // place to determine codec version.
-        );
-    }
-
-    private static String globalMetadataFileName(Metadata metadata) {
-        // 123456789012_test-cluster/cluster-state/dsgYj10Nkso7/global-metadata/metadata__<inverted_metadata_version>__<inverted__timestamp>__<codec_version>
-        return String.join(
-            DELIMITER,
-            METADATA_FILE_PREFIX,
-            RemoteStoreUtils.invertLong(metadata.version()),
-            RemoteStoreUtils.invertLong(System.currentTimeMillis()),
-            String.valueOf(GLOBAL_METADATA_CURRENT_CODEC_VERSION)
-        );
-    }
-
-    private static String metadataAttributeFileName(String componentPrefix, Long metadataVersion) {
-        // 123456789012_test-cluster/cluster-state/dsgYj10Nkso7/global-metadata/<componentPrefix>__<inverted_metadata_version>__<inverted__timestamp>__<codec_version>
-        return String.join(
-            DELIMITER,
-            componentPrefix,
-            RemoteStoreUtils.invertLong(metadataVersion),
-            RemoteStoreUtils.invertLong(System.currentTimeMillis()),
-            String.valueOf(GLOBAL_METADATA_CURRENT_CODEC_VERSION)
-        );
-    }
-
-    private BlobPath getManifestFolderPath(String clusterName, String clusterUUID) {
-        return getCusterMetadataBasePath(clusterName, clusterUUID).add(MANIFEST_PATH_TOKEN);
-    }
-
-    /**
-     * Fetch latest index metadata from remote cluster state
-     *
-     * @param clusterUUID uuid of cluster state to refer to in remote
-     * @param clusterName name of the cluster
-     * @param clusterMetadataManifest manifest file of cluster
-     * @return {@code Map<String, IndexMetadata>} latest IndexUUID to IndexMetadata map
-     */
-    private Map<String, IndexMetadata> getIndexMetadataMap(
-        String clusterName,
-        String clusterUUID,
-        ClusterMetadataManifest clusterMetadataManifest
-    ) {
-        assert Objects.equals(clusterUUID, clusterMetadataManifest.getClusterUUID())
-            : "Corrupt ClusterMetadataManifest found. Cluster UUID mismatch.";
-        Map<String, IndexMetadata> remoteIndexMetadata = new HashMap<>();
-        for (UploadedIndexMetadata uploadedIndexMetadata : clusterMetadataManifest.getIndices()) {
-            IndexMetadata indexMetadata = getIndexMetadata(clusterName, clusterUUID, uploadedIndexMetadata);
-            remoteIndexMetadata.put(uploadedIndexMetadata.getIndexUUID(), indexMetadata);
-        }
-        return remoteIndexMetadata;
-    }
-
-    /**
-     * Fetch index metadata from remote cluster state
-     *
-     * @param clusterUUID uuid of cluster state to refer to in remote
-     * @param clusterName name of the cluster
-     * @param uploadedIndexMetadata {@link UploadedIndexMetadata} contains details about remote location of index metadata
-     * @return {@link IndexMetadata}
-     */
-    private IndexMetadata getIndexMetadata(String clusterName, String clusterUUID, UploadedIndexMetadata uploadedIndexMetadata) {
-        BlobContainer blobContainer = indexMetadataContainer(clusterName, clusterUUID, uploadedIndexMetadata.getIndexUUID());
-        try {
-            String[] splitPath = uploadedIndexMetadata.getUploadedFilename().split("/");
-            return INDEX_METADATA_FORMAT.read(
-                blobContainer,
-                splitPath[splitPath.length - 1],
-                blobStoreRepository.getNamedXContentRegistry()
-            );
-        } catch (IOException e) {
-            throw new IllegalStateException(
-                String.format(Locale.ROOT, "Error while downloading IndexMetadata - %s", uploadedIndexMetadata.getUploadedFilename()),
-                e
-            );
-        }
     }
 
     /**
@@ -1037,7 +584,10 @@ public class RemoteClusterStateService implements Closeable {
      */
     public ClusterState getLatestClusterState(String clusterName, String clusterUUID) {
         start();
-        Optional<ClusterMetadataManifest> clusterMetadataManifest = getLatestClusterMetadataManifest(clusterName, clusterUUID);
+        Optional<ClusterMetadataManifest> clusterMetadataManifest = remoteManifestManager.getLatestClusterMetadataManifest(
+            clusterName,
+            clusterUUID
+        );
         if (clusterMetadataManifest.isEmpty()) {
             throw new IllegalStateException(
                 String.format(Locale.ROOT, "Latest cluster metadata manifest is not present for the provided clusterUUID: %s", clusterUUID)
@@ -1045,10 +595,14 @@ public class RemoteClusterStateService implements Closeable {
         }
 
         // Fetch Global Metadata
-        Metadata globalMetadata = getGlobalMetadata(clusterName, clusterUUID, clusterMetadataManifest.get());
+        Metadata globalMetadata = remoteGlobalMetadataManager.getGlobalMetadata(clusterName, clusterUUID, clusterMetadataManifest.get());
 
         // Fetch Index Metadata
-        Map<String, IndexMetadata> indices = getIndexMetadataMap(clusterName, clusterUUID, clusterMetadataManifest.get());
+        Map<String, IndexMetadata> indices = remoteIndexMetadataManager.getIndexMetadataMap(
+            clusterName,
+            clusterUUID,
+            clusterMetadataManifest.get()
+        );
 
         Map<String, IndexMetadata> indexMetadataMap = new HashMap<>();
         indices.values().forEach(indexMetadata -> { indexMetadataMap.put(indexMetadata.getIndex().getName(), indexMetadata); });
@@ -1057,154 +611,6 @@ public class RemoteClusterStateService implements Closeable {
             .version(clusterMetadataManifest.get().getStateVersion())
             .metadata(Metadata.builder(globalMetadata).indices(indexMetadataMap).build())
             .build();
-    }
-
-    private Metadata getGlobalMetadata(String clusterName, String clusterUUID, ClusterMetadataManifest clusterMetadataManifest) {
-        String globalMetadataFileName = clusterMetadataManifest.getGlobalMetadataFileName();
-        try {
-            // Fetch Global metadata
-            if (globalMetadataFileName != null) {
-                String[] splitPath = globalMetadataFileName.split("/");
-                return GLOBAL_METADATA_FORMAT.read(
-                    globalMetadataContainer(clusterName, clusterUUID),
-                    splitPath[splitPath.length - 1],
-                    blobStoreRepository.getNamedXContentRegistry()
-                );
-            } else if (clusterMetadataManifest.hasMetadataAttributesFiles()) {
-                CoordinationMetadata coordinationMetadata = getCoordinationMetadata(
-                    clusterName,
-                    clusterUUID,
-                    clusterMetadataManifest.getCoordinationMetadata().getUploadedFilename()
-                );
-                Settings settingsMetadata = getSettingsMetadata(
-                    clusterName,
-                    clusterUUID,
-                    clusterMetadataManifest.getSettingsMetadata().getUploadedFilename()
-                );
-                TemplatesMetadata templatesMetadata = getTemplatesMetadata(
-                    clusterName,
-                    clusterUUID,
-                    clusterMetadataManifest.getTemplatesMetadata().getUploadedFilename()
-                );
-                Metadata.Builder builder = new Metadata.Builder();
-                builder.coordinationMetadata(coordinationMetadata);
-                builder.persistentSettings(settingsMetadata);
-                builder.templates(templatesMetadata);
-                clusterMetadataManifest.getCustomMetadataMap()
-                    .forEach(
-                        (key, value) -> builder.putCustom(
-                            key,
-                            getCustomsMetadata(clusterName, clusterUUID, value.getUploadedFilename(), key)
-                        )
-                    );
-                return builder.build();
-            } else {
-                return Metadata.EMPTY_METADATA;
-            }
-        } catch (IOException e) {
-            throw new IllegalStateException(
-                String.format(Locale.ROOT, "Error while downloading Global Metadata - %s", globalMetadataFileName),
-                e
-            );
-        }
-    }
-
-    private CoordinationMetadata getCoordinationMetadata(String clusterName, String clusterUUID, String coordinationMetadataFileName) {
-        try {
-            // Fetch Coordination metadata
-            if (coordinationMetadataFileName != null) {
-                String[] splitPath = coordinationMetadataFileName.split("/");
-                return COORDINATION_METADATA_FORMAT.read(
-                    globalMetadataContainer(clusterName, clusterUUID),
-                    splitPath[splitPath.length - 1],
-                    blobStoreRepository.getNamedXContentRegistry()
-                );
-            } else {
-                return CoordinationMetadata.EMPTY_METADATA;
-            }
-        } catch (IOException e) {
-            throw new IllegalStateException(
-                String.format(Locale.ROOT, "Error while downloading Coordination Metadata - %s", coordinationMetadataFileName),
-                e
-            );
-        }
-    }
-
-    private Settings getSettingsMetadata(String clusterName, String clusterUUID, String settingsMetadataFileName) {
-        try {
-            // Fetch Settings metadata
-            if (settingsMetadataFileName != null) {
-                String[] splitPath = settingsMetadataFileName.split("/");
-                return SETTINGS_METADATA_FORMAT.read(
-                    globalMetadataContainer(clusterName, clusterUUID),
-                    splitPath[splitPath.length - 1],
-                    blobStoreRepository.getNamedXContentRegistry()
-                );
-            } else {
-                return Settings.EMPTY;
-            }
-        } catch (IOException e) {
-            throw new IllegalStateException(
-                String.format(Locale.ROOT, "Error while downloading Settings Metadata - %s", settingsMetadataFileName),
-                e
-            );
-        }
-    }
-
-    private TemplatesMetadata getTemplatesMetadata(String clusterName, String clusterUUID, String templatesMetadataFileName) {
-        try {
-            // Fetch Templates metadata
-            if (templatesMetadataFileName != null) {
-                String[] splitPath = templatesMetadataFileName.split("/");
-                return TEMPLATES_METADATA_FORMAT.read(
-                    globalMetadataContainer(clusterName, clusterUUID),
-                    splitPath[splitPath.length - 1],
-                    blobStoreRepository.getNamedXContentRegistry()
-                );
-            } else {
-                return TemplatesMetadata.EMPTY_METADATA;
-            }
-        } catch (IOException e) {
-            throw new IllegalStateException(
-                String.format(Locale.ROOT, "Error while downloading Templates Metadata - %s", templatesMetadataFileName),
-                e
-            );
-        }
-    }
-
-    private Metadata.Custom getCustomsMetadata(String clusterName, String clusterUUID, String customMetadataFileName, String custom) {
-        requireNonNull(customMetadataFileName);
-        try {
-            // Fetch Custom metadata
-            String[] splitPath = customMetadataFileName.split("/");
-            ChecksumBlobStoreFormat<Metadata.Custom> customChecksumBlobStoreFormat = new ChecksumBlobStoreFormat<>(
-                "custom",
-                METADATA_NAME_FORMAT,
-                (parser -> Metadata.Custom.fromXContent(parser, custom))
-            );
-            return customChecksumBlobStoreFormat.read(
-                globalMetadataContainer(clusterName, clusterUUID),
-                splitPath[splitPath.length - 1],
-                blobStoreRepository.getNamedXContentRegistry()
-            );
-        } catch (IOException e) {
-            throw new IllegalStateException(
-                String.format(Locale.ROOT, "Error while downloading Custom Metadata - %s", customMetadataFileName),
-                e
-            );
-        }
-    }
-
-    /**
-     * Fetch latest ClusterMetadataManifest from remote state store
-     *
-     * @param clusterUUID uuid of cluster state to refer to in remote
-     * @param clusterName name of the cluster
-     * @return ClusterMetadataManifest
-     */
-    public Optional<ClusterMetadataManifest> getLatestClusterMetadataManifest(String clusterName, String clusterUUID) {
-        Optional<String> latestManifestFileName = getLatestManifestFileName(clusterName, clusterUUID);
-        return latestManifestFileName.map(s -> fetchRemoteClusterMetadataManifest(clusterName, clusterUUID, s));
     }
 
     /**
@@ -1216,7 +622,10 @@ public class RemoteClusterStateService implements Closeable {
     public String getLastKnownUUIDFromRemote(String clusterName) {
         try {
             Set<String> clusterUUIDs = getAllClusterUUIDs(clusterName);
-            Map<String, ClusterMetadataManifest> latestManifests = getLatestManifestForAllClusterUUIDs(clusterName, clusterUUIDs);
+            Map<String, ClusterMetadataManifest> latestManifests = remoteManifestManager.getLatestManifestForAllClusterUUIDs(
+                clusterName,
+                clusterUUIDs
+            );
             List<String> validChain = createClusterChain(latestManifests, clusterName);
             if (validChain.isEmpty()) {
                 return ClusterState.UNKNOWN_UUID;
@@ -1231,27 +640,11 @@ public class RemoteClusterStateService implements Closeable {
     }
 
     private Set<String> getAllClusterUUIDs(String clusterName) throws IOException {
-        Map<String, BlobContainer> clusterUUIDMetadata = clusterUUIDContainer(clusterName).children();
+        Map<String, BlobContainer> clusterUUIDMetadata = clusterUUIDContainer(blobStoreRepository, clusterName).children();
         if (clusterUUIDMetadata == null) {
             return Collections.emptySet();
         }
         return Collections.unmodifiableSet(clusterUUIDMetadata.keySet());
-    }
-
-    private Map<String, ClusterMetadataManifest> getLatestManifestForAllClusterUUIDs(String clusterName, Set<String> clusterUUIDs) {
-        Map<String, ClusterMetadataManifest> manifestsByClusterUUID = new HashMap<>();
-        for (String clusterUUID : clusterUUIDs) {
-            try {
-                Optional<ClusterMetadataManifest> manifest = getLatestClusterMetadataManifest(clusterName, clusterUUID);
-                manifest.ifPresent(clusterMetadataManifest -> manifestsByClusterUUID.put(clusterUUID, clusterMetadataManifest));
-            } catch (Exception e) {
-                throw new IllegalStateException(
-                    String.format(Locale.ROOT, "Exception in fetching manifest for clusterUUID: %s", clusterUUID),
-                    e
-                );
-            }
-        }
-        return manifestsByClusterUUID;
     }
 
     /**
@@ -1331,7 +724,7 @@ public class RemoteClusterStateService implements Closeable {
             if (!ClusterState.UNKNOWN_UUID.equals(currentManifest.getPreviousClusterUUID())) {
                 ClusterMetadataManifest previousManifest = trimmedUUIDs.get(currentManifest.getPreviousClusterUUID());
                 if (isMetadataEqual(currentManifest, previousManifest, clusterName)
-                    && isGlobalMetadataEqual(currentManifest, previousManifest, clusterName)) {
+                    && remoteGlobalMetadataManager.isGlobalMetadataEqual(currentManifest, previousManifest, clusterName)) {
                     trimmedUUIDs.remove(clusterUUID);
                 }
             }
@@ -1348,12 +741,20 @@ public class RemoteClusterStateService implements Closeable {
             .stream()
             .collect(Collectors.toMap(md -> md.getIndexName(), Function.identity()));
         for (UploadedIndexMetadata uploadedIndexMetadata : first.getIndices()) {
-            final IndexMetadata firstIndexMetadata = getIndexMetadata(clusterName, first.getClusterUUID(), uploadedIndexMetadata);
+            final IndexMetadata firstIndexMetadata = remoteIndexMetadataManager.getIndexMetadata(
+                clusterName,
+                first.getClusterUUID(),
+                uploadedIndexMetadata
+            );
             final UploadedIndexMetadata secondUploadedIndexMetadata = secondIndices.get(uploadedIndexMetadata.getIndexName());
             if (secondUploadedIndexMetadata == null) {
                 return false;
             }
-            final IndexMetadata secondIndexMetadata = getIndexMetadata(clusterName, second.getClusterUUID(), secondUploadedIndexMetadata);
+            final IndexMetadata secondIndexMetadata = remoteIndexMetadataManager.getIndexMetadata(
+                clusterName,
+                second.getClusterUUID(),
+                secondUploadedIndexMetadata
+            );
             if (firstIndexMetadata.equals(secondIndexMetadata) == false) {
                 return false;
             }
@@ -1361,123 +762,12 @@ public class RemoteClusterStateService implements Closeable {
         return true;
     }
 
-    private boolean isGlobalMetadataEqual(ClusterMetadataManifest first, ClusterMetadataManifest second, String clusterName) {
-        Metadata secondGlobalMetadata = getGlobalMetadata(clusterName, second.getClusterUUID(), second);
-        Metadata firstGlobalMetadata = getGlobalMetadata(clusterName, first.getClusterUUID(), first);
-        return Metadata.isGlobalResourcesMetadataEquals(firstGlobalMetadata, secondGlobalMetadata);
-    }
-
     private boolean isValidClusterUUID(ClusterMetadataManifest manifest) {
         return manifest.isClusterUUIDCommitted();
     }
 
-    /**
-     * Fetch ClusterMetadataManifest files from remote state store in order
-     *
-     * @param clusterUUID uuid of cluster state to refer to in remote
-     * @param clusterName name of the cluster
-     * @param limit max no of files to fetch
-     * @return all manifest file names
-     */
-    private List<BlobMetadata> getManifestFileNames(String clusterName, String clusterUUID, int limit) throws IllegalStateException {
-        try {
-
-            /*
-              {@link BlobContainer#listBlobsByPrefixInSortedOrder} will list the latest manifest file first
-              as the manifest file name generated via {@link RemoteClusterStateService#getManifestFileName} ensures
-              when sorted in LEXICOGRAPHIC order the latest uploaded manifest file comes on top.
-             */
-            return manifestContainer(clusterName, clusterUUID).listBlobsByPrefixInSortedOrder(
-                MANIFEST_FILE_PREFIX + DELIMITER,
-                limit,
-                BlobContainer.BlobNameSortOrder.LEXICOGRAPHIC
-            );
-        } catch (IOException e) {
-            throw new IllegalStateException("Error while fetching latest manifest file for remote cluster state", e);
-        }
-    }
-
-    /**
-     * Fetch latest ClusterMetadataManifest file from remote state store
-     *
-     * @param clusterUUID uuid of cluster state to refer to in remote
-     * @param clusterName name of the cluster
-     * @return latest ClusterMetadataManifest filename
-     */
-    private Optional<String> getLatestManifestFileName(String clusterName, String clusterUUID) throws IllegalStateException {
-        List<BlobMetadata> manifestFilesMetadata = getManifestFileNames(clusterName, clusterUUID, 1);
-        if (manifestFilesMetadata != null && !manifestFilesMetadata.isEmpty()) {
-            return Optional.of(manifestFilesMetadata.get(0).name());
-        }
-        logger.info("No manifest file present in remote store for cluster name: {}, cluster UUID: {}", clusterName, clusterUUID);
-        return Optional.empty();
-    }
-
-    /**
-     * Fetch ClusterMetadataManifest from remote state store
-     *
-     * @param clusterUUID uuid of cluster state to refer to in remote
-     * @param clusterName name of the cluster
-     * @return ClusterMetadataManifest
-     */
-    private ClusterMetadataManifest fetchRemoteClusterMetadataManifest(String clusterName, String clusterUUID, String filename)
-        throws IllegalStateException {
-        try {
-            return getClusterMetadataManifestBlobStoreFormat(filename).read(
-                manifestContainer(clusterName, clusterUUID),
-                filename,
-                blobStoreRepository.getNamedXContentRegistry()
-            );
-        } catch (IOException e) {
-            throw new IllegalStateException(String.format(Locale.ROOT, "Error while downloading cluster metadata - %s", filename), e);
-        }
-    }
-
-    private ChecksumBlobStoreFormat<ClusterMetadataManifest> getClusterMetadataManifestBlobStoreFormat(String fileName) {
-        long codecVersion = getManifestCodecVersion(fileName);
-        if (codecVersion == MANIFEST_CURRENT_CODEC_VERSION) {
-            return CLUSTER_METADATA_MANIFEST_FORMAT;
-        } else if (codecVersion == ClusterMetadataManifest.CODEC_V1) {
-            return CLUSTER_METADATA_MANIFEST_FORMAT_V1;
-        } else if (codecVersion == ClusterMetadataManifest.CODEC_V0) {
-            return CLUSTER_METADATA_MANIFEST_FORMAT_V0;
-        }
-
-        throw new IllegalArgumentException("Cluster metadata manifest file is corrupted, don't have valid codec version");
-    }
-
-    private int getManifestCodecVersion(String fileName) {
-        String[] splitName = fileName.split(DELIMITER);
-        if (splitName.length == SPLITED_MANIFEST_FILE_LENGTH) {
-            return Integer.parseInt(splitName[splitName.length - 1]); // Last value would be codec version.
-        } else if (splitName.length < SPLITED_MANIFEST_FILE_LENGTH) { // Where codec is not part of file name, i.e. default codec version 0
-                                                                      // is used.
-            return ClusterMetadataManifest.CODEC_V0;
-        } else {
-            throw new IllegalArgumentException("Manifest file name is corrupted");
-        }
-    }
-
-    public static String encodeString(String content) {
-        return Base64.getUrlEncoder().withoutPadding().encodeToString(content.getBytes(StandardCharsets.UTF_8));
-    }
-
     public void writeMetadataFailed() {
         getStats().stateFailed();
-    }
-
-    /**
-     * Exception for Remote state transfer.
-     */
-    static class RemoteStateTransferException extends RuntimeException {
-
-        public RemoteStateTransferException(String errorDesc) {
-            super(errorDesc);
-        }
-
-        public RemoteStateTransferException(String errorDesc, Throwable cause) {
-            super(errorDesc, cause);
-        }
     }
 
     /**
@@ -1490,7 +780,7 @@ public class RemoteClusterStateService implements Closeable {
         clusterUUIDs.forEach(clusterUUID -> {
             getBlobStoreTransferService().deleteAsync(
                 ThreadPool.Names.REMOTE_PURGE,
-                getCusterMetadataBasePath(clusterName, clusterUUID),
+                getCusterMetadataBasePath(blobStoreRepository, clusterName, clusterUUID),
                 new ActionListener<>() {
                     @Override
                     public void onResponse(Void unused) {
@@ -1529,7 +819,7 @@ public class RemoteClusterStateService implements Closeable {
         try {
             getBlobStoreTransferService().listAllInSortedOrderAsync(
                 ThreadPool.Names.REMOTE_PURGE,
-                getManifestFolderPath(clusterName, clusterUUID),
+                remoteManifestManager.getManifestFolderPath(clusterName, clusterUUID),
                 "manifest",
                 Integer.MAX_VALUE,
                 new ActionListener<>() {
@@ -1576,7 +866,7 @@ public class RemoteClusterStateService implements Closeable {
             Set<String> staleIndexMetadataPaths = new HashSet<>();
             Set<String> staleGlobalMetadataPaths = new HashSet<>();
             activeManifestBlobMetadata.forEach(blobMetadata -> {
-                ClusterMetadataManifest clusterMetadataManifest = fetchRemoteClusterMetadataManifest(
+                ClusterMetadataManifest clusterMetadataManifest = remoteManifestManager.fetchRemoteClusterMetadataManifest(
                     clusterName,
                     clusterUUID,
                     blobMetadata.name()
@@ -1594,7 +884,7 @@ public class RemoteClusterStateService implements Closeable {
                 }
             });
             staleManifestBlobMetadata.forEach(blobMetadata -> {
-                ClusterMetadataManifest clusterMetadataManifest = fetchRemoteClusterMetadataManifest(
+                ClusterMetadataManifest clusterMetadataManifest = remoteManifestManager.fetchRemoteClusterMetadataManifest(
                     clusterName,
                     clusterUUID,
                     blobMetadata.name()
@@ -1681,7 +971,7 @@ public class RemoteClusterStateService implements Closeable {
 
     private void deleteStalePaths(String clusterName, String clusterUUID, List<String> stalePaths) throws IOException {
         logger.debug(String.format(Locale.ROOT, "Deleting stale files from remote - %s", stalePaths));
-        getBlobStoreTransferService().deleteBlobs(getCusterMetadataBasePath(clusterName, clusterUUID), stalePaths);
+        getBlobStoreTransferService().deleteBlobs(getCusterMetadataBasePath(blobStoreRepository, clusterName, clusterUUID), stalePaths);
     }
 
     /**
@@ -1709,35 +999,5 @@ public class RemoteClusterStateService implements Closeable {
 
     public RemotePersistenceStats getStats() {
         return remoteStateStats;
-    }
-
-    private static class UploadedMetadataResults {
-        List<UploadedIndexMetadata> uploadedIndexMetadata;
-        Map<String, UploadedMetadataAttribute> uploadedCustomMetadataMap;
-        UploadedMetadataAttribute uploadedCoordinationMetadata;
-        UploadedMetadataAttribute uploadedSettingsMetadata;
-        UploadedMetadataAttribute uploadedTemplatesMetadata;
-
-        public UploadedMetadataResults(
-            List<UploadedIndexMetadata> uploadedIndexMetadata,
-            Map<String, UploadedMetadataAttribute> uploadedCustomMetadataMap,
-            UploadedMetadataAttribute uploadedCoordinationMetadata,
-            UploadedMetadataAttribute uploadedSettingsMetadata,
-            UploadedMetadataAttribute uploadedTemplatesMetadata
-        ) {
-            this.uploadedIndexMetadata = uploadedIndexMetadata;
-            this.uploadedCustomMetadataMap = uploadedCustomMetadataMap;
-            this.uploadedCoordinationMetadata = uploadedCoordinationMetadata;
-            this.uploadedSettingsMetadata = uploadedSettingsMetadata;
-            this.uploadedTemplatesMetadata = uploadedTemplatesMetadata;
-        }
-
-        public UploadedMetadataResults() {
-            this.uploadedIndexMetadata = new ArrayList<>();
-            this.uploadedCustomMetadataMap = new HashMap<>();
-            this.uploadedCoordinationMetadata = null;
-            this.uploadedSettingsMetadata = null;
-            this.uploadedTemplatesMetadata = null;
-        }
     }
 }
